+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/VPSResetProgressModel.java
@@ -66,6 +66,11 @@ public class VPSResetProgressModel extends ProgressModel<GameRepresentation> {
     if (games.size() > MAX_REFRESH_COUNT) {
       EventManager.getInstance().notifyTablesChanged();
     }
+    else {
+      for (GameRepresentation game : games) {
+        EventManager.getInstance().notifyTableChange(game.getId(), null);
+      }
+    }
   }
 
   @Override
@@ -73,9 +78,6 @@ public class VPSResetProgressModel extends ProgressModel<GameRepresentation> {
     try {
       game.setVpsUpdates(new VPSChanges());
       client.getGameService().saveGame(game);
-      if (games.size() <= MAX_REFRESH_COUNT) {
-        EventManager.getInstance().notifyTableChange(game.getId(), null);
-      }
     }
     catch (Exception e) {
       LOG.error("Failed to reset VPS indicator: " + e.getMessage(), e);