+++ RELEASE_NOTES.md
@@ -23,6 +23,7 @@
 - **Tables / PUP Packs**: Added [PupPackScreenTweaker.exe](https://github.com/matiou11/PupPackScreenTweaker) as additional PUP pack editor to the PUP pack section. Note that this editor is only available when working on the cabinet itself.
 - **Tables / Table Data Manager**: The auto-naming has no restrictions on VPX files that are located in sub-folders anymore. You can also rename them now.
 - **Tables / Backglasses**: Added button for DMD positioning.
+- **Tables / VPS Section**: This section supports multi-selection now. This way, you can bulk auto-match a selection of tables. 
 - **Tables / Highscores Section**: This section has undergone a revamp:
   - Added support for multi-selection from the table overview.
   - Added bulk operation support for highscore resets.